// Cursado/Practico/TrabajosPracticos/Evaluables/TP6/DeliveryEat/src/app/models/datosCompartidos.ts
export class DatosCompartidos {
    numeroTarjeta: string = "";
    nombreTarjeta: string = "";
    cvTarjeta: string = "";
    vencimientoTarjeta: string = "";

    direccion: string = "";
    numeroDireccion: string = "";
    ciudad: string = ""
    referenciaOpcional: string = "";

    metodoPagoVisa: boolean = false;
    montoApagar: string = "";
    vuelto: string = "";

    fechaSeleccionada: boolean = false;
    fecha: string = "";
    horaEnvio: string = "";
    minutosEnvio: string = "";


    DatosCompartidos() {

    }
}